Move the certificate mode to seperate file
