+++ pkg/cable/libreswan/libreswan.go
@@ -59,8 +59,6 @@ const (
 	ikeportArg       = "--ikeport"
 	dpdactionHoldArg = "--dpdaction=hold"
 	dpddelayArg      = "--dpddelay"
-	certArg          = "--cert"
-	caArg            = "--ca"
 )
 
 // AuthMode defines the authentication mode for libreswan.
@@ -108,9 +106,8 @@ type libreswan struct {
 	plutoStarted          bool
 	authMode              AuthMode
 
-	stopCh                chan struct{}
-	certificateController *CertificateHandler
-	signingRequestor      certificate.SigningRequestor
+	certificateHandler *CertificateHandler
+	signingRequestor   certificate.SigningRequestor
 }
 
 type specification struct {
@@ -137,12 +134,6 @@ func NewLibreswan(localEndpoint *submendpoint.Local, _ *types.SubmarinerCluster,
 
 	// Parse and validate authentication mode with fallback to PSK
 	authModeStr := ipSecSpec.AuthMode
-	if authModeStr == "" {
-		// Fallback to default PSK mode if not set
-		authModeStr = string(AuthModePSK)
-
-		logger.Info("CE_IPSEC_AUTHMODE not set, defaulting to psk authentication")
-	}
 
 	authMode := AuthMode(authModeStr)
 	if authMode != AuthModePSK && authMode != AuthModeCert {
@@ -227,27 +218,23 @@ func (i *libreswan) Init() error {
 
 		fmt.Fprintf(file, "%%any %%any : PSK \"%s\"\n", i.secretKey)
 	} else if i.authMode == AuthModeCert {
-		logger.Info("Setting up certificate authentication")
-
-		logger.Infof("Starting certificate creation with Private IPs %s Public IPs %s", i.localEndpoint.PrivateIPs, i.localEndpoint.PublicIPs)
+		logger.Infof("Issuing certificate with Private IPs %s Public IPs %s", i.localEndpoint.PrivateIPs, i.localEndpoint.PublicIPs)
 		sanIPs := make([]string, 0, len(i.localEndpoint.PrivateIPs)+len(i.localEndpoint.PublicIPs))
 		sanIPs = append(sanIPs, i.localEndpoint.PrivateIPs...)
 		sanIPs = append(sanIPs, i.localEndpoint.PublicIPs...)
 
-		i.stopCh = make(chan struct{})
-
 		var err error
 
-		i.certificateController = NewCertificateHandler(i.localEndpoint.ClusterID)
+		i.certificateHandler = NewCertificateHandler(i.localEndpoint.ClusterID)
 
-		err = i.signingRequestor.Issue(context.TODO(), "submariner-certificate", sanIPs, i.certificateController.OnSignedCallback)
+		err = i.signingRequestor.Issue(context.TODO(), "libreswan-"+i.localEndpoint.Hostname, sanIPs, i.certificateHandler.OnSignedCallback)
 		if err != nil {
 			logger.Warningf("Unable to issue certificate: %v", err)
 		}
 
-		logger.Info("Started certificate syncers and controller")
-	} else {
-		return fmt.Errorf("unsupported authentication mode: %s", i.authMode)
+		i.plutoStarted = true
+
+		logger.Info("Successfully issued certificate")
 	}
 
 	return nil
@@ -431,156 +418,6 @@ func (i *libreswan) ConnectToEndpoint(endpointInfo *natdiscovery.NATEndpointInfo
 	}
 
 	return i.connectToEndpointPSKMode(endpointInfo)
-}
-
-func appendConnectionStanza(confPath, stanza, connName string) error {
-	if err := removeConnectionStanza(confPath, connName); err != nil {
-		return err
-	}
-
-	f, err := os.OpenFile(confPath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0o644)
-	if err != nil {
-		return errors.Wrap(err, "error opening file")
-	}
-
-	defer f.Close()
-
-	if !strings.HasSuffix(stanza, "\n") {
-		stanza += "\n"
-	}
-
-	_, err = f.WriteString(stanza)
-
-	return errors.Wrap(err, "error writing to file")
-}
-
-func removeConnectionStanza(confPath, connName string) error {
-	data, err := os.ReadFile(confPath)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return nil
-		}
-
-		return errors.Wrap(err, "error reading file")
-	}
-
-	lines := strings.Split(string(data), "\n")
-	var out []string
-	inStanza := false
-
-	for _, line := range lines {
-		if strings.HasPrefix(line, "conn ") && strings.TrimSpace(line) == ("conn "+connName) {
-			inStanza = true
-			continue
-		}
-
-		if inStanza && strings.HasPrefix(line, "conn ") && strings.TrimSpace(line) != ("conn "+connName) {
-			inStanza = false
-		}
-
-		if !inStanza {
-			out = append(out, line)
-		}
-	}
-
-	for len(out) > 0 && out[len(out)-1] == "" {
-		out = out[:len(out)-1]
-	}
-
-	if len(out) == 0 {
-		return errors.Wrap(os.Remove(confPath), "error removing file")
-	}
-
-	return errors.Wrap(os.WriteFile(confPath, []byte(strings.Join(out, "\n")+"\n"), 0o600), "error writing file")
-}
-
-func (i *libreswan) connectToEndpointCertMode(endpointInfo *natdiscovery.NATEndpointInfo) (string, error) {
-	logger.Info("Certificate mode: skipping pluto start and whack; assuming pluto is managed externally and config/NSS DB are updated.")
-
-	confPath := SubmarinerConfPath()
-	endpoint := &endpointInfo.Endpoint
-	leftID := "submariner-client-" + i.localEndpoint.ClusterID
-	left := i.localEndpoint.GetPrivateIP(endpointInfo.UseFamily)
-	right := endpointInfo.UseIP
-
-	leftSubnets := i.localEndpoint.ExtractSubnetsExcludingIP(endpointInfo.UseIP)
-	rightSubnets := endpoint.Spec.ExtractSubnetsExcludingIP(endpointInfo.UseIP)
-
-	for lsi, leftSubnet := range leftSubnets {
-		for rsi, rightSubnet := range rightSubnets {
-			connName := toConnectionName(endpoint.Spec.CableName, endpointInfo.UseFamily, lsi, rsi)
-
-			encapsulationLine := ""
-			if endpointInfo.UseNAT || i.forceUDPEncapsulation {
-				encapsulationLine = "    encapsulation=yes\n"
-			}
-
-			conf := fmt.Sprintf(`conn %s
-    left=%s
-    leftid=%%fromcert
-    leftcert=%s
-    leftrsasigkey=%%cert
-    leftsubnet=%s
-    leftmodecfgclient=false
-    right=%s
-    rightid=%%fromcert
-    rightsubnet=%s
-%s    auto=add
-    ikev2=insist
-    authby=rsasig
-    type=tunnel`,
-				connName,
-				left,
-				leftID,
-				leftSubnet,
-				right,
-				rightSubnet,
-				encapsulationLine,
-			)
-			if err := appendConnectionStanza(confPath, conf, connName); err != nil {
-				logger.Errorf(err, "Failed to append connection stanza to %s", confPath)
-				return "", err
-			}
-
-			logger.Infof("Appended Libreswan connection config for %s to %s", connName, confPath)
-
-			output, err := command.New(exec.Command("ipsec", "auto", "--add", connName)).CombinedOutput()
-			if err != nil {
-				logger.Errorf(err, "Failed to add connection with ipsec auto --add: %s", string(output))
-				return "", errors.Wrap(err, "failed to add connection")
-			} else {
-				logger.Infof("Added connection with ipsec auto --add: %s", string(output))
-			}
-
-			connectionMode := i.calculateOperationMode(&endpoint.Spec)
-
-			logger.Infof("Connection mode for %s: %v", connName, connectionMode)
-
-			if connectionMode == operationModeClient || connectionMode == operationModeBidirectional {
-				whackArgs := []string{"--name", connName, "--initiate"}
-				//nolint:gosec // ipsec whack args are from trusted config
-				output, err = command.New(exec.Command("ipsec", append([]string{"whack"}, whackArgs...)...)).CombinedOutput()
-				if err != nil {
-					logger.Errorf(err, "Failed to bring up connection %s with whack: %s", connName, string(output))
-					return "", errors.Wrap(err, "failed to bring up connection")
-				} else {
-					logger.Infof("Brought up connection %s with whack: %s", connName, string(output))
-				}
-			}
-		}
-	}
-
-	i.connections = append(i.connections,
-		subv1.Connection{
-			Endpoint: endpoint.Spec,
-			UsingIP:  endpointInfo.UseIP,
-			UsingNAT: endpointInfo.UseNAT,
-			Status:   subv1.Connected,
-		})
-
-	i.plutoStarted = true
-
-	return endpointInfo.UseIP, nil
 }
 
 func (i *libreswan) connectToEndpointPSKMode(endpointInfo *natdiscovery.NATEndpointInfo) (string, error) {
@@ -815,16 +652,6 @@ func (i *libreswan) DisconnectFromEndpoint(endpoint *types.SubmarinerEndpoint, f
 	return nil
 }
 
-func (i *libreswan) disconnectCertMode(connectionName string) error {
-	if err := removeConnectionStanza(SubmarinerConfPath(), connectionName); err != nil {
-		return errors.Wrapf(err, "failed to remove connection stanza for %q", connectionName)
-	}
-
-	logger.Infof("Removed Libreswan connection config for %q", connectionName)
-
-	return nil
-}
-
 func (i *libreswan) disconnectPSKMode(connectionName string) error {
 	args := []string{"--delete", nameArg, connectionName}
 
@@ -935,12 +762,8 @@ func (i *libreswan) waitForControlSocket() error {
 func (i *libreswan) Cleanup() error {
 	logger.Info("Uninstalling the libreswan cable driver")
 
-	if i.certificateController != nil {
-		i.certificateController.Cleanup()
-	}
-
-	if i.stopCh != nil {
-		close(i.stopCh)
+	if i.certificateHandler != nil {
+		i.certificateHandler.Cleanup()
 	}
 
 	// Delete submariner.conf on uninstall
